requirements.txt
fastapi==0.110.3
uvicorn==0.29.0
python-multipart==0.0.9
Jinja2==3.1.3
torch==2.3.1
torchvision==0.18.1
Pillow==10.3.0
